♻️ useFade 간소화
